feat(cmd): simpler flux cmd
A new flux cmd is added to internal as a proposal to replace the
existing one. The CLI is more similar to Python where no parameter is
specified and it goes to the REPL or a single parameter is given for
running a script.

The script is inferred to be a file and it can be changed to an inline
script using an option. This is more similar to other interpreted
language command lines which makes bash completion easier for filenames
than the `@file` syntax.

The non-REPL method of executing a script now matches with how influxdb
would execute the script. This avoids a difference when testing
execution by allowing the CLI to use the same code path rather than to
hijack the REPL path which can produce different results.